requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
httpx
paramiko
gspread
google-auth
groq
